Refactored Fixtures Changed int to Integer in Ratings Added more tests to RecommenderAPI with tests

+++ test/models/RatingTest.java
@@ -25,7 +25,7 @@ public class RatingTest
 		Rating rating = new Rating(1, 10, 5, 234324234);
 		assertEquals(1, rating.userId);
 		assertEquals(10, rating.movieId);
-		assertEquals(5, rating.rating);
+		assertEquals((Integer) 5, rating.rating);
 		assertEquals(234324234, rating.timestamp);
 	}
